# perf(ext/web): add op_encode_binary_string